+++ src/components/dashboard/website-management/services/services-management.jsx
@@ -8,6 +8,7 @@ import styles from "./services-mgmt.module.css";
 import AddServiceModal from "./add-service-modal";
 import EditServiceModal from "./edit-service-modal";
 import { openBootstrapModal } from "../../../../utility/open-bootstrap-modal";
+import EditAndDeleteButtons from "../../edit-and-delete-btns/edit-and-delete-btns";
 
 const ServicesManagement = () => {
   const [state, dispatch] = useStore();
@@ -18,21 +19,22 @@ const ServicesManagement = () => {
   }
 
   const deleteService = async (serviceId) => {
-    const response = await httpDELETE(
-      api.clinic_services.delete_service + serviceId
-    );
-    if (response.status === 400) {
-      const data = await response.json();
-      console.log(data);
-      return;
+    if (window.confirm("هل تريد الحذف فعلاً؟") == true) {
+      const response = await httpDELETE(
+        api.clinic_services.delete_service + serviceId
+      );
+      if (response.status === 400) {
+        alert("Error!!");
+        return;
+      }
+      dispatch("DELETE_CLINIC_SERVICE", serviceId);
     }
-    dispatch("DELETE_CLINIC_SERVICE", serviceId);
   };
   const [serviceToEdit, setServiceToEdit] = useState({});
 
   return (
     <div className="container">
-      <div className="card text-center">
+      <div className="card text-center m-3">
         <div className="card-header">الخيارات المتاحة</div>
         <div className="card-body">
           <h5 className="card-title mt-4">هذه الإجراءات خاصة بخدمات العيادة</h5>
@@ -40,9 +42,9 @@ const ServicesManagement = () => {
             يمكنك الإختيار ما بين إضافة خدة أو إستعراض الخدمات الحالية أو أو حتى
             حذف خدمة.
           </p>
-          {/* <hr /> */}
-          <div className="d-flex justify-content-around">
-            <div className="col-4">
+
+          <div className="row m-0">
+            <div className="col-md-4 col-sm-8 offset-md-4 offset-sm-2 d-flex justify-content-center">
               <ShowModalButton
                 color="blue"
                 modalId="#addServiceModal"
@@ -54,7 +56,7 @@ const ServicesManagement = () => {
         <div className="card-footer text-muted">لا تنسى أن تحظى بيوم سعيد</div>
       </div>
 
-      <div className="row justify-content-between">
+      <div className="row m-3 justify-content-between">
         {state.clinic_services.length !== 0 && (
           <h1 className="text-center text-white mt-3">الخدمات الحالية</h1>
         )}
@@ -66,27 +68,15 @@ const ServicesManagement = () => {
         {state.clinic_services.map((servc) => (
           <div
             key={servc.id}
-            className="col-5 bg-white position-relative  border-success border-opacity-50 rounded-3 m-2 p-1"
+            className="col-sm-12 col-lg-5 bg-white position-relative  border-success border-opacity-50 rounded-3 p-1 my-1"
           >
-            <svg
-              className={styles.deleteButton}
-              onClick={() => deleteService(servc.id)}
-              viewBox="0 0 16 16"
-            >
-              <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM5.354 4.646a.5.5 0 1 0-.708.708L7.293 8l-2.647 2.646a.5.5 0 0 0 .708.708L8 8.707l2.646 2.647a.5.5 0 0 0 .708-.708L8.707 8l2.647-2.646a.5.5 0 0 0-.708-.708L8 7.293 5.354 4.646z" />
-            </svg>
+            <div className="top-0 end-0 position-absolute">
 
-            <svg
-              className={styles.editButton}
-              onClick={() => {
-                setServiceToEdit(servc);
-                openBootstrapModal("showEditServiceModelBtn");
-              }}
-              viewBox="0 0 16 16"
-            >
-              <path d="M12.854.146a.5.5 0 0 0-.707 0L10.5 1.793 14.207 5.5l1.647-1.646a.5.5 0 0 0 0-.708l-3-3zm.646 6.061L9.793 2.5 3.293 9H3.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.207l6.5-6.5zm-7.468 7.468A.5.5 0 0 1 6 13.5V13h-.5a.5.5 0 0 1-.5-.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.5-.5V10h-.5a.499.499 0 0 1-.175-.032l-.179.178a.5.5 0 0 0-.11.168l-2 5a.5.5 0 0 0 .65.65l5-2a.5.5 0 0 0 .168-.11l.178-.178z" />
-            </svg>
-
+            <EditAndDeleteButtons deleteAction={() => deleteService(servc.id)} editAction={() => {
+              setServiceToEdit(servc);
+              openBootstrapModal("showEditServiceModelBtn");
+            }} />
+            </div>
             {servc.image && (
               <img
                 src={api.base_url + servc.image}
@@ -102,11 +92,11 @@ const ServicesManagement = () => {
               />
             )}
             <h4 className="p-2 mb-0">{servc.title}</h4>
-            <p className="p-1 mb-5">{servc.text} </p>
+            <p className="p-1 mb-5 text-truncate">{servc.text} </p>
           </div>
         ))}
         <AddServiceModal />
-        <EditServiceModal service={serviceToEdit}/>
+        <EditServiceModal service={serviceToEdit} />
       </div>
     </div>
   );
